feat(docker-compose): use a separate config file for internal nets
This was causing errors such as the following in docker-compose:

```
network rabbitmq declared as external, but could not be found
```

This is because `external` means that the network should be managed
outside the lifecycle of the application and must be created first. This
seems like the incorrect move in this case.