fix(docker-compose.sh): load the env file, fix shellcheck
docker-compose, somewhat confusingly, has two overloaded meanings for
"env file":
1. Expanded inside the container, by passing environment variables to
   the started process inside the container, which is
   services.SVC.env_file
2. Expanded inside docker-compose files, which is --env-file

We were missing both of them in different places, and the latter is
fixed by this commit.

https://docs.docker.com/compose/environment-variables/set-environment-variables/

Fixes these warnings:
WARN[0000] The "NOMOS_DB_USER" variable is not set. Defaulting to a blank string.
WARN[0000] The "NOMOS_DB_PASSWORD" variable is not set. Defaulting to a blank string.
WARN[0000] The "NOMOS_RABBITMQ_USER" variable is not set. Defaulting to a blank string.
WARN[0000] The "NOMOS_RABBITMQ_PASSWORD" variable is not set. Defaulting to a blank string.
WARN[0000] The "NOMOS_RABBITMQ_VHOST" variable is not set. Defaulting to a blank string.