tidy: delete unused pin and rfid *login* code
This is all dead code: the ability to log in with pin or rfid was ripped
out from the interface in 2015 in
e44d9f6f9c38c078daa9a54b3873d178213629e6 but nobody removed the
implementation.

I cannot conceive of any use case where being able to *log in* with
either rfid or pin to get your user's privileges is a good idea for
security reasons, so we should just remove the code altogether. All the
reasonable use cases for these are better served with what was
actually built: CheckPin and CheckRfid, executed by a service account on the
machine that needs to check pins/rfid credentials and do something based
on them.