Make a working docker-compose dev env (#297)
* fix(docker-compose.sh): load the env file, fix shellcheck

docker-compose, somewhat confusingly, has two overloaded meanings for
"env file":
1. Expanded inside the container, by passing environment variables to
   the started process inside the container, which is
   services.SVC.env_file
2. Expanded inside docker-compose files, which is --env-file

We were missing both of them in different places, and the latter is
fixed by this commit.

https://docs.docker.com/compose/environment-variables/set-environment-variables/

Fixes these warnings:
WARN[0000] The "NOMOS_DB_USER" variable is not set. Defaulting to a blank string.
WARN[0000] The "NOMOS_DB_PASSWORD" variable is not set. Defaulting to a blank string.
WARN[0000] The "NOMOS_RABBITMQ_USER" variable is not set. Defaulting to a blank string.
WARN[0000] The "NOMOS_RABBITMQ_PASSWORD" variable is not set. Defaulting to a blank string.
WARN[0000] The "NOMOS_RABBITMQ_VHOST" variable is not set. Defaulting to a blank string.

* feat(docker-compose): add self-contained dev configuration

This seems to be the most useful configuration to use for development so
we should ship it for ease of use.

* fix(docker-compose): load the environment into the webhooker service

Previously, the webhooker service was always pulling defaults and was
not actually using any of its variables in the environment file.

* docs: update repo URL in README

* fix(docker-compose): make the mysql service actually work

I found several bugs:
1. The service was taking 30 seconds (!!!?!) to start on my machine
   (Arch Linux x86_64) and taking 16 GB of memory (...). This seems to
   be some kind of strange bug in the container, which is worked around
   by setting some ulimit values.
2. MYSQL_PASS is the wrong variable for the password; the correct one is
   MYSQL_PASSWORD.
3. A root password must be set for the container to start.
4. We forgot to set the database to create, which is important to giving
   the app access to it.

* chore(webhooker): npm added name to the lockfile

* feat(setup): add a script to create an API key for the webhook service

Previously this was done in the tools/vagrant_provision.sh script, but
I want to deprecate Vagrant, so we need a second way.

* chore: populate the env template with plausible dev values

We were also missing some environment variables related to Stripe which
the backend was angry about, so I fixed those too.

* docs: move docker-compose.md where it will be seen, write up new devenv

* feat(docker-compose): use a separate config file for internal nets

This was causing errors such as the following in docker-compose:

```
network rabbitmq declared as external, but could not be found
```

This is because `external` means that the network should be managed
outside the lifecycle of the application and must be created first. This
seems like the incorrect move in this case.

---------

Co-authored-by: TyIsI <TyIsI@users.noreply.github.com>